Fixed a memory corruption issue.

+++ include/sockaddr.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-struct sockaddr *create_sockaddr(char *address, uint16_t port, size_t **sockaddr_length);
+struct sockaddr *create_sockaddr(char *address, uint16_t port, unsigned int **sockaddr_length);
 int destroy_sockaddr(struct sockaddr *sockaddr);
 
 #endif
